validate file exist
Former-commit-id: 86387ceaac2be39bf3cccd30335dfb7fb46dc271

+++ func.php
@@ -368,19 +368,21 @@ function getRandomItemFromArray($array)
  */
 function removeDuplicateLines($filePath)
 {
-  // Read the file into an array, each line as an element
-  $lines = file($filePath, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
+  if (file_exists($filePath)) {
+    // Read the file into an array, each line as an element
+    $lines = file($filePath, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
 
-  if (is_array($lines)) {
-    // Remove duplicate lines
-    $lines = array_unique($lines);
+    if (is_array($lines)) {
+      // Remove duplicate lines
+      $lines = array_unique($lines);
 
-    // Remove empty strings from the array
-    $lines = array_filter($lines, function ($value) {
-      return trim($value) !== '';
-    });
-    // Write the modified lines back to the file
-    file_put_contents($filePath, implode("\n", $lines) . "\n");
+      // Remove empty strings from the array
+      $lines = array_filter($lines, function ($value) {
+        return trim($value) !== '';
+      });
+      // Write the modified lines back to the file
+      file_put_contents($filePath, implode("\n", $lines) . "\n");
+    }
   }
 }
